# docs/api/openapi.yaml
components:
  schemas:
    api.SwaggerAddMemberRequest:
      properties:
        role:
          example: member
          type: string
        user_id:
          example: 874d6314-6353-45e9-ab2a-5fe930ea4dbc
          type: string
      type: object
...
    api.SwaggerCreateSpaceRequest:
      properties:
        description:
          example: A project space
          type: string
        icon:
          example: rocket
          type: string
        is_private:
          example: false
          type: boolean
        name:
          example: My Project
          type: string
        slug:
          example: my-project
          type: string
        type:
          example: project
          type: string
      type: object
...
    api.SwaggerErrorDetail:
      properties:
        code:
          example: UNAUTHORIZED
          type: string
        message:
          example: invalid email or password
          type: string
        request_id:
          example: req_a8d9912a
          type: string
      type: object
    api.SwaggerErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/api.SwaggerErrorDetail'
      type: object
...
    api.SwaggerUpdateOrgRequest:
      properties:
        description:
          example: Updated description
          type: string
        name:
          example: Updated Org
          type: string
      type: object
...
    api.SwaggerUpdateSpaceRequest:
      properties:
        description:
          example: Updated description
          type: string
        icon:
          example: star
          type: string
        is_private:
          example: false
          type: boolean
        name:
          example: Updated Name
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: 'JWT Bearer token. Format: "Bearer <token>". Obtain via POST /api/v1/auth/login'
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: hello@azimuthalhq.com
    name: Azimuthal HQ
    url: https://azimuthalhq.com
  description: |-
    The fully open-source alternative to Jira, Confluence, and Jira Service Desk.
    Self-hostable, single binary, Apache 2.0 licensed.

    ## Authentication
    All endpoints except /auth/login and /auth/register require a Bearer JWT token.
    Obtain a token via POST /api/v1/auth/login, then include it as:
    `Authorization: Bearer <your-token>`
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Azimuthal API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /orgs/{orgID}:
    get:
      description: Returns an organization by its ID.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Organization details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid org ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
      security:
      - BearerAuth: []
      summary: Get organization
      tags:
      - spaces
    patch:
      description: Updates an organization's name and description (preserves plan).
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerUpdateOrgRequest'
                description: Updated org fields
                summary: body
        description: Updated org fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Updated organization
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - spaces
...
  /orgs/{orgID}/spaces:
    get:
      description: Returns all spaces belonging to the specified organization.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  additionalProperties: {}
                  type: object
                type: array
          description: List of spaces
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid org ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: List spaces
      tags:
      - spaces
    post:
      description: Creates a new space in the organization. Type must be 'project',
        'wiki', or 'service_desk'.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerCreateSpaceRequest'
                description: Space details
                summary: body
        description: Space details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Created space
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Create space
      tags:
      - spaces
  /orgs/{orgID}/spaces/{spaceID}:
    delete:
      description: Soft-deletes a space by ID.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Delete space
      tags:
      - spaces
    get:
      description: Returns a single space by ID.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Space details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
      security:
      - BearerAuth: []
      summary: Get space
      tags:
      - spaces
    put:
      description: Updates a space's name, description, icon, and privacy setting.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerUpdateSpaceRequest'
                description: Updated fields
                summary: body
        description: Updated fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Updated space
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Update space
      tags:
      - spaces
...
  /orgs/{orgID}/spaces/{spaceID}/members:
    get:
      description: Returns all members of the specified space.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  additionalProperties: {}
                  type: object
                type: array
          description: List of members
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: List space members
      tags:
      - members
    post:
      description: Adds a user as a member of the specified space.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerAddMemberRequest'
                description: Member details
                summary: body
        description: Member details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Member added
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Add space member
      tags:
      - members
  /orgs/{orgID}/spaces/{spaceID}/members/{userID}:
    delete:
      description: Removes a user from the specified space.
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: orgID
        required: true
        schema:
          type: string
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: User ID (UUID)
        in: path
        name: userID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Removed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Remove space member
      tags:
      - members
...
servers:
- url: localhost:8080/api/v1